pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "email-validator>=2.3.0",
    "fastapi>=0.121.0",
    "firebase-admin>=7.1.0",
    "gunicorn>=23.0.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=7.0.1",
    "stripe>=13.2.0",
    "supabase>=2.24.0",
    "uvicorn>=0.38.0",
]
